## refactor : 입력받은 문자열을 매개변수로 사칙연산을 수행하는 메서드 구현 및 불필요한 메서드 삭제
